requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
matplotlib==3.9.0
google-generativeai==0.7.2
python-dotenv==1.0.1
